Added authorization to correct paths, added ideas and author to user and idea response objects

+++ src/idea/idea.service.ts
@@ -3,62 +3,58 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 
 import { IdeaEntity } from './idea.entity';
-import { IdeaDTO } from './idea.dto';
+import { IdeaDTO, IdeaRO } from './idea.dto';
+import { UserEntity } from '../user/user.entity';
 
 @Injectable()
 export class IdeaService {
   constructor(
     @InjectRepository(IdeaEntity)
-    private ideaRepository:Repository<IdeaEntity>
-  ) {}
+    private ideaRepository: Repository<IdeaEntity>,
+    @InjectRepository(UserEntity)
+    private userRepository: Repository<UserEntity>
+  ) { }
 
-  async showAll() {
-    return await this.ideaRepository.find();
+  async showAll(): Promise<Array<IdeaRO>> {
+    const ideas = await this.ideaRepository.find({ relations: ['author'] });
+    return ideas.map(idea => idea.toResponseObject());
   }
 
-  async create(data:IdeaDTO) {
-    data = this.pick(data, Object.keys(IdeaDTO));
-    const idea = await this.ideaRepository.create(data);
+  async create(userId: string, data: IdeaDTO): Promise<IdeaRO> {
+    const user = await this.userRepository.findOne({ where: { id: userId } });
+    const idea = await this.ideaRepository.create({ ...data, author: user });
     await this.ideaRepository.save(idea);
-    return idea;
+    return idea.toResponseObject();
   }
 
-  async read(id:string) {
-    const idea = await this.ideaRepository.findOne({ where: { id } });
+  async read(id: string): Promise<IdeaRO> {
+    const idea = await this.ideaRepository.findOne({ where: { id }, relations: ['author'] });
     if (!idea) throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
-    return idea;
+    return idea.toResponseObject();
   }
 
-  async update(id:string, data:Partial<IdeaDTO>) {
-    data = this.pick(data, Object.keys(IdeaDTO));
-    let idea = await this.ideaRepository.findOne({ where: { id } });
+  async update(id: string, userId: string, data: Partial<IdeaDTO>): Promise<IdeaRO> {
+    let idea = await this.ideaRepository.findOne({ where: { id }, relations: ['author'] });
     if (!idea) throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
-    idea = await this.ideaRepository.save({...idea, ...data});
-    return idea;
+    this.ensureOwnership(idea, userId);
+    idea = await this.ideaRepository.save({ ...idea, ...data });
+    return idea.toResponseObject();
   }
 
-  async destroy(id:string) {
-    const idea = await this.ideaRepository.findOne({ where: { id } });
+  async destroy(id: string, userId: string): Promise<IdeaRO> {
+    const idea = await this.ideaRepository.findOne({ where: { id }, relations: ['author'] });
     if (!idea) throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+    this.ensureOwnership(idea, userId);
     await this.ideaRepository.delete({ id });
-    return idea;
+    return idea.toResponseObject();
   }
 
 
+  /* UTILITY METHODS */
 
-  /* Utility Methods */
-
-  /**
-   * Picks specific paths from object
-   * @param  obj   Object to pick from
-   * @param  paths Paths picked
-   * @return       Cleaned object
-   */
-  private pick(obj, paths) {
-    Object.keys(obj)
-      .filter(key => !paths.includes(key))
-      .forEach(key => delete obj[key]);
-    return obj;
+  private ensureOwnership(idea: IdeaEntity, userId: string): void {
+    if (idea.author.id !== userId)
+      throw new HttpException('Incorrect user', HttpStatus.UNAUTHORIZED);
   }
 
 }
